drone/requirements.txt
opencv-python
paho-mqtt
gradio